Выполнено задание 5.12 Модуляция
